+++ packages/layouts/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.10.0](https://github.com/ttoss/ttoss/compare/@ttoss/layouts@0.9.5...@ttoss/layouts@0.10.0) (2026-02-15)
+
+### Features
+
+- react wizard ([#870](https://github.com/ttoss/ttoss/issues/870)) ([d309e78](https://github.com/ttoss/ttoss/commit/d309e78f798064c4e4e9a57cef94872c40f94d15))
+
 ## [0.9.5](https://github.com/ttoss/ttoss/compare/@ttoss/layouts@0.9.4...@ttoss/layouts@0.9.5) (2026-02-03)
 
 **Note:** Version bump only for package @ttoss/layouts